requirements.txt
pytest==7.3.0
selenium==4.8.3
webdriver-manager==4.0.0
Faker==18.4.0
requests==2.28.2
allure-pytest==2.13.2
